Fix clean up after github-related tests execution; remove ssh-key of github.com

+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/git/PullRequestPluginTest.java
@@ -117,9 +117,12 @@ public class PullRequestPluginTest {
     List<String> listPullRequest =
         gitHubClientService.getNumbersOfOpenedPullRequests(
             NAME_REPO, gitHubUsername, gitHubPassword);
-    gitHubClientService.closePullRequest(
-        NAME_REPO, Collections.max(listPullRequest), gitHubUsername, gitHubPassword);
-    gitHubClientService.deleteBranch(NAME_REPO, NEW_BRANCH, gitHubUsername, gitHubPassword);
+
+    if (!listPullRequest.isEmpty()) {
+      gitHubClientService.closePullRequest(
+          NAME_REPO, Collections.max(listPullRequest), gitHubUsername, gitHubPassword);
+      gitHubClientService.deleteBranch(NAME_REPO, NEW_BRANCH, gitHubUsername, gitHubPassword);
+    }
   }
 
   @Test(priority = 0)
